refactor: 중복되는 의미 제거, pathParameter 네이밍 명확하게 수정
