+++ src/test/java/com/dinosaur/foodbowl/IntegrationTest.java
@@ -24,7 +24,6 @@ import com.dinosaur.foodbowl.domain.comment.application.CommentFindService;
 import com.dinosaur.foodbowl.domain.comment.application.CommentService;
 import com.dinosaur.foodbowl.domain.comment.dao.CommentRepository;
 import com.dinosaur.foodbowl.domain.follow.application.FollowService;
-import com.dinosaur.foodbowl.domain.follow.application.GetFollowService;
 import com.dinosaur.foodbowl.domain.follow.dao.FollowRepository;
 import com.dinosaur.foodbowl.domain.post.PostTestHelper;
 import com.dinosaur.foodbowl.domain.post.application.PostFindService;
@@ -134,9 +133,6 @@ public class IntegrationTest {
 
   @SpyBean
   protected FollowService followService;
-
-  @SpyBean
-  protected GetFollowService getFollowService;
 
   @SpyBean
   protected PostService postService;
