+++ src/test/java/com/dinosaur/foodbowl/domain/follow/application/FollowServiceTest.java
@@ -1,12 +1,19 @@
 package com.dinosaur.foodbowl.domain.follow.application;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.dinosaur.foodbowl.IntegrationTest;
+import com.dinosaur.foodbowl.domain.follow.dto.FollowerResponseDto;
+import com.dinosaur.foodbowl.domain.follow.dto.FollowingResponseDto;
 import com.dinosaur.foodbowl.domain.user.UserTestHelper.UserBuilder;
 import com.dinosaur.foodbowl.domain.user.entity.User;
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 
 class FollowServiceTest extends IntegrationTest {
 
@@ -89,6 +96,49 @@ class FollowServiceTest extends IntegrationTest {
       followService.unfollow(me, other);
       em.flush();
       em.clear();
+    }
+
+    @Test
+    @DisplayName("유저를 팔로우한 유저 목록 조회를 성공한다.")
+    public void getFollowers() {
+      User me = userTestHelper.builder().build();
+      User follower1 = userTestHelper.builder().build();
+      User follower2 = userTestHelper.builder().build();
+      User following1 = userTestHelper.builder().build();
+
+      followService.follow(follower1, me);
+      followService.follow(follower2, me);
+      followService.follow(me, follower1);
+      followService.follow(me, following1);
+
+      Pageable pageable = PageRequest.of(0, 5);
+
+      List<FollowerResponseDto> response = followService.getFollowers(me, pageable);
+      List<Long> responseUserIds = response.stream().map(FollowerResponseDto::getUserId).toList();
+
+      assertThat(response.size()).isEqualTo(2);
+      assertThat(responseUserIds).contains(follower1.getId(), follower2.getId());
+    }
+
+    @Test
+    @DisplayName("유저가 팔로우한 유저 목록 조회를 성공한다.")
+    public void getFollowings() {
+      User me = userTestHelper.builder().build();
+      User follower1 = userTestHelper.builder().build();
+      User following1 = userTestHelper.builder().build();
+      User following2 = userTestHelper.builder().build();
+
+      followService.follow(me, following1);
+      followService.follow(me, following2);
+      followService.follow(follower1, me);
+      followService.follow(following1, me);
+
+      Pageable pageable = PageRequest.of(0, 5);
+      List<FollowingResponseDto> response = followService.getFollowings(me, pageable);
+      List<Long> responseUserIds = response.stream().map(FollowingResponseDto::getUserId).toList();
+
+      assertThat(response.size()).isEqualTo(2);
+      assertThat(responseUserIds).contains(following1.getId(), following2.getId());
     }
   }
 }
